Greatly improved float parsing and the like

+++ src/code/canhandler.h
@@ -18,10 +18,10 @@ class CanHandler : public QObject
 
 
     Q_PROPERTY(qint64 rpmData READ rpmData WRITE setRpmData NOTIFY rpmDataChanged)
-    Q_PROPERTY(qint64 coolantData READ coolantData WRITE setCoolantData NOTIFY coolantDataChanged)
-    Q_PROPERTY(qint64 afrData READ afrData WRITE setAfrData NOTIFY afrDataChanged)
-    Q_PROPERTY(qint64 biasData READ biasData WRITE setBiasData NOTIFY biasDataChanged)
-    Q_PROPERTY(qint64 voltageData READ voltageData WRITE setVoltageData NOTIFY voltageDataChanged)
+    Q_PROPERTY(float coolantData READ coolantData WRITE setCoolantData NOTIFY coolantDataChanged)
+    Q_PROPERTY(float afrData READ afrData WRITE setAfrData NOTIFY afrDataChanged)
+    Q_PROPERTY(float biasData READ biasData WRITE setBiasData NOTIFY biasDataChanged)
+    Q_PROPERTY(float voltageData READ voltageData WRITE setVoltageData NOTIFY voltageDataChanged)
 
     QML_ELEMENT
 
@@ -46,14 +46,14 @@ public:
 
     void setRpmData(const qint64 &a);
     qint64 rpmData() const;
-    void setCoolantData(const qint64 &a);
-    qint64 coolantData() const;
-    void setAfrData(const qint64 &a);
-    qint64 afrData() const;
-    void setBiasData(const qint64 &a);
-    qint64 biasData() const;
-    void setVoltageData(const qint64 &a);
-    qint64 voltageData() const;
+    void setCoolantData(const float &a);
+    float coolantData() const;
+    void setAfrData(const float &a);
+    float afrData() const;
+    void setBiasData(const float &a);
+    float biasData() const;
+    void setVoltageData(const float &a);
+    float voltageData() const;
 
     void setSocketCanStatus(const QString &a);
     QString socketCanStatus() const;
@@ -80,10 +80,10 @@ private:
     qint64 m_canTestData = -1;
 
     qint64 m_rmpData = -1;
-    qint64 m_coolantData = -1;
-    qint64 m_afrData = -1;
-    qint64 m_biasData = -1;
-    qint64 m_voltageData = -1;
+    float m_coolantData = -1;
+    float m_afrData = -1;
+    float m_biasData = -1;
+    float m_voltageData = -1;
 
     std::unique_ptr<QCanBusDevice> m_canDevice;
     QTimer *m_busStatusTimer = nullptr;
